pyproject.toml
[tool.poetry]
name = "fan_controller"
version = "0.1.0"
description = ""
authors = ["Anton Tinyakov <anlorn@anlorn.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"RPi.GPIO" = "^0.7.1"
Flask = "^2.2.2"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "fan_controller.main:main"
